+++ src/components/Root.tsx
@@ -1,6 +1,7 @@
 import { useEffect, useRef, useState } from 'react';
 import { ModalProvider } from '@react-aria/overlays';
 import { StyleSheetManager } from 'styled-components';
+import { useViewportSize } from '@react-aria/utils';
 
 import {
   BASE_STYLES,
@@ -45,6 +46,9 @@ export interface CubeRootProps extends BaseProps {
   applyLegacyTokens?: boolean;
 }
 
+const IS_DVH_SUPPORTED =
+  typeof CSS?.supports !== 'undefined' ? CSS.supports('height: 100dvh') : false;
+
 export function Root(allProps: CubeRootProps) {
   let {
     children,
@@ -63,6 +67,44 @@ export function Root(allProps: CubeRootProps) {
 
   const [rootRef, setRootRef] = useState();
 
+  // We need to measure the window's height in JS rather than using percentages in CSS
+  // so that contents (e.g. menu) can inherit the max-height properly. Using percentages
+  // does not work properly because there is nothing to base the percentage on.
+  // We cannot use vh units because mobile browsers adjust the window height dynamically
+  // when the address bar/bottom toolbars show and hide on scroll and vh units are fixed.
+  // Also, the visual viewport is smaller than the layout viewport when the virtual keyboard
+  // is up, so use the VisualViewport API to ensure the tray is displayed above the keyboard.
+  let viewport = useViewportSize();
+  let [height, setHeight] = useState(
+    IS_DVH_SUPPORTED ? undefined : viewport.height,
+  );
+  let timeoutRef = useRef<NodeJS.Timeout>();
+
+  useEffect(() => {
+    if (IS_DVH_SUPPORTED) {
+      return;
+    }
+
+    if (timeoutRef.current) {
+      clearTimeout(timeoutRef.current);
+    }
+
+    // When the height is decreasing, and the keyboard is visible
+    // (visual viewport smaller than layout viewport), delay setting
+    // the new max height until after the animation is complete
+    // so that there isn't an empty space under the tray briefly.
+    if (
+      viewport.height < (height || 0) &&
+      viewport.height < window.innerHeight
+    ) {
+      timeoutRef.current = setTimeout(() => {
+        setHeight(viewport.height);
+      }, 500);
+    } else {
+      setHeight(viewport.height);
+    }
+  }, [height, viewport.height]);
+
   useEffect(() => {
     if (!rootRef) {
       // @ts-ignore
@@ -79,6 +121,9 @@ export function Root(allProps: CubeRootProps) {
           ref={ref}
           {...filterBaseProps(props, { eventProps: true })}
           styles={styles}
+          style={{
+            '--cube-dynamic-viewport-height': height ? height + 'px' : '100dvh',
+          }}
         >
           <GlobalStyles
             bodyStyles={bodyStyles}
