+++ src/components/overlays/Modal/Tray.tsx
@@ -1,6 +1,5 @@
 import { useDOMRef } from '@react-spectrum/utils';
-import { useViewportSize } from '@react-aria/utils';
-import { forwardRef, useEffect, useRef, useState } from 'react';
+import { forwardRef } from 'react';
 import { useModal, useOverlay, usePreventScroll } from '@react-aria/overlays';
 
 import { BaseProps, Props, Styles, tasty } from '../../../tasty';
@@ -24,7 +23,7 @@ const TrayWrapperElement = tasty({
 const TrayElement = tasty({
   styles: {
     zIndex: 2,
-    height: 'max (@cube-visual-viewport-height * .9)',
+    height: 'max 90dvh',
     width: '288px 90vw',
     pointerEvents: 'auto',
     transition:
@@ -104,39 +103,6 @@ let TrayWrapper = forwardRef(function TrayWrapper(
     isDisabled: isNonModal,
   });
 
-  // We need to measure the window's height in JS rather than using percentages in CSS
-  // so that contents (e.g. menu) can inherit the max-height properly. Using percentages
-  // does not work properly because there is nothing to base the percentage on.
-  // We cannot use vh units because mobile browsers adjust the window height dynamically
-  // when the address bar/bottom toolbars show and hide on scroll and vh units are fixed.
-  // Also, the visual viewport is smaller than the layout viewport when the virtual keyboard
-  // is up, so use the VisualViewport API to ensure the tray is displayed above the keyboard.
-  let viewport = useViewportSize();
-  let [height, setHeight] = useState(viewport.height);
-  let timeoutRef = useRef<NodeJS.Timeout>();
-
-  useEffect(() => {
-    if (timeoutRef.current) {
-      clearTimeout(timeoutRef.current);
-    }
-
-    // When the height is decreasing, and the keyboard is visible
-    // (visual viewport smaller than layout viewport), delay setting
-    // the new max height until after the animation is complete
-    // so that there isn't an empty space under the tray briefly.
-    if (viewport.height < height && viewport.height < window.innerHeight) {
-      timeoutRef.current = setTimeout(() => {
-        setHeight(viewport.height);
-      }, 500);
-    } else {
-      setHeight(viewport.height);
-    }
-  }, [height, viewport.height]);
-
-  let wrapperStyle = {
-    '--cube-visual-viewport-height': height + 'px',
-  };
-
   let domProps = mergeProps(otherProps, overlayProps);
 
   return (
@@ -144,7 +110,6 @@ let TrayWrapper = forwardRef(function TrayWrapper(
       mods={{
         open: isOpen,
       }}
-      style={wrapperStyle}
     >
       <TrayElement
         qa={qa || 'Tray'}
